Divination Card Update
Individual Divination Cards are now viewable in their entirety, showing composed images reflecting exactly how they look in game.

Added a small javascript function to limit font-size when flavour text overflows outside the bounds of the card.

Wrote a small barebones web scraper in order to fetch data from the wiki and update the back-end accordingly.
This was used to grab most of the cards' flavour texts and will most likely be used again in the future to add any other necessary data when possible.